- Move HttpClient initialization logic to separate class - Remove DI from RestGithubApiClientIntegrationTest
